feat(institution): 모바일 대응
